add virtual ImporterContext.PreprocessAsync
