+++ Assets/UniGLTF/Runtime/UniGLTF/IO/RenameUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ded08d75fb22a94f80265d4403d5b3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
