+++ Assets/VRM10/Runtime/IO/Vrm10Exporter.cs
@@ -235,9 +235,8 @@ namespace UniVRM10
             }
 
             // Fix Duplicated name
-            gltfExporter.FixName(Storage.Gltf);
+            RenameUtil.FixName(Storage.Gltf);
         }
-
 
         /// <summary>
         /// VRMコンポーネントのエクスポート
